Celiac pathfinder
From original concept document: ISP 605 -  Michelle Waldenmaier - CELIAC Disease Pathfinder
	
"This pathfinder is targeted at the local (NY Capital District) library patron that wishes to learn more about Celiac disease in order to lead a gluten-free lifestyle. Celiac disease is sensitivity to Gluten, a protein found in wheat, barley, rye, some oat products, and several other grains. Celiac disease interferes with the absorption of vital nutrients for the body causing an immunological toxic reaction and can be the trigger to other debilitating conditions. There is no known cure for Celiac disease, the only treatment is an absolutely 100% gluten-free diet. Resources given in this pathfinder are useful to the newly diagnosed patient as well as the veteran patient whom is maintaining the gluten-free lifestyle. While these sources are comprehensive in context, they are by no means exhaustive in the resources available to the celiac patient. However, there should be some critical regard to resources that are targeted at selling a product to the celiac patient."